+++ demo/+config.ts
@@ -5,14 +5,4 @@ import docpress from '@brillout/docpress/config'
 
 const config = {
   extends: docpress,
-  choices: {
-    server: {
-      choices: ['Hono', 'Express', 'Fastify'],
-      default: 'Hono',
-    },
-    uiFrameworkVikeExtension: {
-      choices: ['vike-react', 'vike-vue', 'vike-solid'],
-      default: 'vike-react',
-    },
-  },
 } satisfies Config